requirements.txt
# Wireless channel coding and equalization experiment dependencies
numpy>=1.23.0
scipy>=1.9.0
matplotlib>=3.6.0
pytest>=7.0.0
pylint>=2.15.0
python-docx>=1.1.0
python-pptx>=0.6.23
